# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Realona Exchange API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades:
    get:
      operationId: getTrades
      tags: [trades]
      summary: Get current user trades (as buyer or seller)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
    post:
      operationId: createTrade
      tags: [trades]
      summary: Initiate a trade (buy a listing)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
  /trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      summary: Get trade details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
  /trades/{id}/confirm-payment:
    post:
      operationId: confirmTradePayment
      tags: [trades]
      summary: Buyer confirms payment (escrow funded)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
  /trades/{id}/seller-transferred:
    post:
      operationId: sellerTransferred
      tags: [trades]
      summary: Seller marks account as transferred
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
  /trades/{id}/confirm-receipt:
    post:
      operationId: confirmReceipt
      tags: [trades]
      summary: Buyer confirms receipt - releases funds to seller
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
  /trades/{id}/dispute:
    post:
      operationId: openDispute
      tags: [trades]
      summary: Open a dispute on a trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"

  # ─── CHAT ──────────────────────────────────────────────────────────────────
...
  /trades/{id}/cancel:
    post:
      operationId: cancelTrade
      tags: [trades]
      summary: Cancel a pending trade (buyer or seller)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Trade cancelled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessResponse" }

  # ─── TRADE CREDENTIALS ─────────────────────────────────────────────────────
  /trades/{id}/credentials:
    get:
      operationId: getTradeCredentials
      tags: [trades]
      summary: Get revealed credentials for a trade (buyer only, after payment confirmed)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TradeCredentials" }

  # ─── CONFIRM ACCESS ────────────────────────────────────────────────────────
  /trades/{id}/confirm-access:
    post:
      operationId: confirmTradeAccess
      tags: [trades]
      summary: Buyer confirms they have full access to the account
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Access confirmed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessResponse" }

  # ─── TRADE RATINGS ─────────────────────────────────────────────────────────
  /trades/{id}/rate:
    post:
      operationId: rateTradePartner
      tags: [trades]
      summary: Submit a star rating for the trade partner after completion
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TradeRatingInput" }
      responses:
        "200":
          description: Rating submitted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessResponse" }

  # ─── OFFERS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]

    # ── Auth ──
...
    Trade:
      type: object
      required: [id, listingId, buyerId, sellerId, amount, status, accessConfirmed, createdAt]
      properties:
        id:
          type: integer
        listingId:
          type: integer
        buyerId:
          type: integer
        sellerId:
          type: integer
        buyerUsername:
          type: ["string", "null"]
        sellerUsername:
          type: ["string", "null"]
        gameName:
          type: ["string", "null"]
        pictureUrl:
          type: ["string", "null"]
        listingCategory:
          type: ["string", "null"]
        amount:
          type: number
        agreedAmount:
          type: ["number", "null"]
        fee:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, payment_confirmed, seller_transferred, completed, disputed, refunded, cancelled]
        accessConfirmed:
          type: boolean
        disputeReason:
          type: ["string", "null"]
        buyerRating:
          type: ["number", "null"]
        sellerRating:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TradeInput:
      type: object
      required: [listingId]
      properties:
        listingId:
          type: integer
    DisputeInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string

    # ── Trade Messages ──
...
    TradeCredentials:
      type: object
      properties:
        konamiId:
          type: ["string", "null"]
        konamiPassword:
          type: ["string", "null"]
        accessCode:
          type: ["string", "null"]
        accountEmail:
          type: ["string", "null"]
        accountPassword:
          type: ["string", "null"]
    TradeRatingInput:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
...